Added methods for working with API (part 3)

+++ app/src/main/java/com/example/afishaapp/data/module/search/DateRange.kt
@@ -1,6 +1,6 @@
 package com.example.afishaapp.data.module.search
 
 data class DateRange(
-    val start: Int,
-    val end: Int
+    val start: Long,
+    val end: Long
 )
